requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
boto3==1.18.48
botocore==1.21.48
Django==3.2.7
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-ses==2.2.2
future==0.18.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
jmespath==0.10.0
Pillow==8.3.2
promise==2.3
psycopg2==2.9.1
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2021.1
Rx==1.6.1
s3transfer==0.5.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.7